+++ src/app/api/gameInfo/route.ts
@@ -1,0 +1,10 @@
+import { NextResponse } from "next/server";
+
+const { PrismaClient } = require("@prisma/client");
+
+const prisma = new PrismaClient();
+
+export async function GET(request: Request) {
+  const games = await prisma.game_INFO.findMany();
+  return NextResponse.json({ games });
+}
